feat: Make project standalone Zephyr module
The original idea of this fork seems to be to be integrated directly
into the Zephyr project. (see following RFC [1] and pull request [2]).

This has been denied with the proposal to make it a standalone Zephyr
module that doesn't require direct dependencies into the Zephyr root
project.

This patch fulfills this idea using the original PR [2] by removing the
external Kconfig and CMake dependencies and shipping these two files
with it.

[1]: https://github.com/zephyrproject-rtos/zephyr/issues/61822
[2]: https://github.com/zephyrproject-rtos/zephyr/pull/61821

Signed-off-by: Sophie 'Tyalie' Friedrich <dev@flowerpot.me>